// marine/sse/src/generated/tlsc-sse-reports-schema.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type TheSitenameSchema = string;
export type TheSitenumberSchema = number;
export type TheSiteTypeSchema = "FIXED" | "FLOATING";
export type TheLastUpdateSchema = string;
export type TheSeastateSchema = "CALM" | "LIGHT" | "MODERATE" | "BREEZE" | "GALE" | "STORM";
export type TheTrendSchema = "UNKNOWN" | "NO_CHANGE" | "ASCENDING" | "DESCENDING";
export type TheWindwavedirSchema = number;
export type TheConfidenceSchema = "POOR" | "MODERATE" | "GOOD";
export type TheTemperatureSchema = number;
export type TheHeelAngleSchema = number;
export type TheLightStatusSchema = "ON" | "OFF" | "ON_D";
export type TheCoordLatitudeSchema = number;
export type TheCoordLongitudeSchema = number;
export type TheSseReportsSchema = TheItemsSchema[];

export interface TheSSEReportRootSchema {
  SSE_Reports: TheSseReportsSchema;
}
export interface TheItemsSchema {
  Site: TheSiteSchema;
  SSE_Fields: TheSseFieldsSchema;
  Extra_Fields?: TheExtraFieldsSchema;
  [k: string]: unknown;
}
export interface TheSiteSchema {
  SiteName: TheSitenameSchema;
  SiteNumber: TheSitenumberSchema;
  SiteType: TheSiteTypeSchema;
  [k: string]: unknown;
}
export interface TheSseFieldsSchema {
  Last_Update: TheLastUpdateSchema;
  SeaState: TheSeastateSchema;
  Trend: TheTrendSchema;
  WindWaveDir: TheWindwavedirSchema;
  Confidence: TheConfidenceSchema;
  [k: string]: unknown;
}
export interface TheExtraFieldsSchema {
  Temperature?: TheTemperatureSchema;
  Heel_Angle?: TheHeelAngleSchema;
  Light_Status?: TheLightStatusSchema;
  Coord_Latitude?: TheCoordLatitudeSchema;
  Coord_Longitude?: TheCoordLongitudeSchema;
  [k: string]: unknown;
}